+++ settings.gradle
@@ -24,8 +24,7 @@ include 'eventmesh-starter'
 include 'eventmesh-examples'
 include 'eventmesh-spi'
 include 'eventmesh-openconnect:eventmesh-openconnect-java'
-include 'eventmesh-connectors:source-connector-rocketmq'
-include 'eventmesh-connectors:sink-connector-rocketmq'
+include 'eventmesh-connectors:eventmesh-connector-rocketmq'
 include 'eventmesh-connectors:eventmesh-openfunction-connector'
 include 'eventmesh-storage-plugin:eventmesh-storage-api'
 include 'eventmesh-storage-plugin:eventmesh-storage-standalone'
